% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter_biv.R
\name{enter_biv}
\alias{enter_biv}
\title{Enter summary bivariate statistics}
\usage{
enter_biv()
}
\value{
Returns an object
}
\description{
You will be prompted for all information
}
\details{
You will be prompted for a character string that describes the source
of the data, an integer giving the sample size, a real value for the mean of x,
a real value for the mean of y, the variance of x, the variance of y, and the
covariance of x and y.  Note that the x variable is the "explanatory" variable
and y is the "explained value."  In calibration terminology, x is the known value
from a "gold standard" and y is the calibration measurement.  In biological
anthropology x could be known stature and y could be a bone length (as in allometry)
or x could be known age and y could be osteon population density.
}
\examples{
\dontrun{
enter_biv() # There are no parameters. You will be prompted for everything.
}
}
