% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{femur_train}
\alias{femur_train}
\title{Femur data}
\format{
A data frame with 37 rows and 6 variables
}
\source{
{
See "European data set: Excel" and "Notes" under "Special Projects"
\url{https://fae.johnshopkins.edu/chris-ruff/}
}
}
\usage{
data(femur_train)
}
\description{
A dataset of Italian females.  The data come from Ruff (2018).  Statures were
estimated using modifications of the Fully method (Niskanen & Ruff 2018; Raxter et
al. 2006, 2007).
}
\details{
Data as follows:
\itemize{
 \item{Stature:}{  Stature in mm.}
 \item{Fem:}{  Maximum femur length in mm.}
}
}
\references{
Niskanen, M. and Ruff, C.B. (2018) Body size and shape reconstruction. In Ruff CB, editor.
\emph{Skeletal Variation and Adaptation in Europeans: Upper Paleolithic to the Twentieth Century}.
Hoboken: Wiley-Blackwell, pp. 15-37.

Raxter, M.H., Auerbach, B.M. and Ruff, C.B. (2006) A revision of the Fully technique
for estimating statures. \emph{Am. J. Phys. Anthropol}. 130, 374-384.

Raxter, M.H., Ruff, C.B. and Auerbach, B.M. (2007) Technical note: Revised Fully
stature estimation technique. \emph{Am. J. Phys. Anthropol}. 133, 817-818.

Ruff, C.B. (ed.) (2018) \emph{Skeletal variation and adaptation in Europeans:
Upper Paleolithic to the Twentieth Century}. Hoboken: Wiley Blackwell.
}
\keyword{datasets}
