% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HL_Table_1}
\alias{HL_Table_1}
\title{Example Table 1 from Hunter & Lamboy (1981)}
\format{
A data frame with 20 rows and two variables
}
\usage{
data(HL_Table_1)
}
\description{
A fake (?) dataset of "Known" molybdenum (x variable) and measured
molybdenum (y variable).
}
\details{
Data as follows:
\itemize{
 \item{Known_molyb:}{  "Known" molybdenum}
 \item{Measured_molyb:}{  Measured molybdenum}
}
}
\references{
Hunter, William G., and Warren F. Lamboy (1981)	A Bayesian analysis of
the linear calibration problem. \emph{Technometrics} 23:323-350.
}
\keyword{datasets}
